requirements.txt
asgiref==3.4.1
attrs==21.2.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.4.0
greenlet==1.1.0
tornado==4.5.3
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
PyMySQL==1.0.2
python-dotenv==0.18.0
requests==2.25.1
SQLAlchemy==1.4.20
cryptography==3.4.8
imageio==2.9.0
matplotlib==3.4.3
pytz
boto3